pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["pyorerun"]

[project]
name = "pyorerun"
version = " 1.5.0"
authors = [{name = "Pierre Puchaud", email = "puchaud.pierre@gmail.com"}]
maintainers = [{name = "Pierre Puchaud", email = "puchaud.pierre@gmail.com"}]
description = "A Python package to rerun C3D files and biomechanical simulations."
license = {file = "LICENSE.md"}
readme = "README.md"
requires-python = ">=3.11"
classifiers = [
         "Programming Language :: Python :: 3",
         "License :: OSI Approved :: MIT License",
         "Operating System :: OS Independent",
     ]
dependencies = [
    "ezc3d",  # Now needed, but imported for our custom PyoMarkers.from_c3d()
    "numpy",
    "rerun-sdk==0.24.1",
    "trimesh",
    "tk",
    "imageio",
    "imageio-ffmpeg",
    "matplotlib",
#    "opensim", # Not yet available on pypi, use `conda install opensim-org opensim=4.5.1`
#    "biorbd" # Not yet available on pypi, use `conda install -c conda-forge biorbd`
    ]
keywords = ["c3d", "motion capture", "rerun", "biorbd", "markers"]

[project.urls]
homepage = "http://github.com/Ipuch/pyorerun"
